pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "expmate"
version = "0.1.3"
description = "ML Research Boilerplate — Config & Logging First"
readme = "README.md"
license = {text = "MIT"}
authors = [
  {name = "Kunhee Kim", email = "kunhee.kim@kaist.ac.kr"}
]
keywords = [
  "machine-learning",
  "deep-learning",
  "experiment-tracking",
  "pytorch",
  "configuration",
  "logging",
  "reproducibility",
  "research",
  "cli"
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Intended Audience :: Developers",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
requires-python = ">=3.8"
dependencies = [
  "numpy>=1.21.0",
  "matplotlib>=3.5.0",
  "polars>=0.20.0",
  "psutil>=5.9.0",
  "pyyaml>=6.0",
]

[project.urls]
Homepage = "https://github.com/kunheek/expmate"
Repository = "https://github.com/kunheek/expmate"
Documentation = "https://github.com/kunheek/expmate#readme"
Changelog = "https://github.com/kunheek/expmate/blob/main/CHANGELOG.md"
Issues = "https://github.com/kunheek/expmate/issues"

[project.scripts]
expmate = "expmate.cli:main"

[project.optional-dependencies]
wandb = ["wandb>=0.15.0"]
tensorboard = ["tensorboard>=2.13.0"]
tracking = ["wandb>=0.15.0", "tensorboard>=2.13.0"]
dev = [
  "pytest>=7.0.0",
  "pytest-cov>=4.0.0",
  "black>=23.0.0",
  "ruff>=0.1.0",
  "mypy>=1.0.0",
  "pre-commit>=3.0.0"
]

[tool.setuptools.packages.find]
where = ["src"]
